effective-apache-airflow/requirements.txt
pip~=20.1.1
setuptools~=49.2.0
apache-airflow~=1.10.10